Split returns wrong value for long separator strings
If CBA_fnc_split is asked to split an input string with a separator
string longer than the source string, the function will return an
array of ASCII values of the input string. The correct behavior
is to return the input string. Fixes #133.

Added a few more useful test cases for CBA_fnc_split.