Restore equivalent position handling in taskDefend
Previously if no position was given the position of the group was used instead.

When I updated the function to use `params` and then retroactively check if the default position value was used I forget to make the `_position` variable use the *position* of the group rather than just the group itself.