Restore fnc_searchNearby leader behaviour
In actuality this is something of a compromise.

Having the leader enter the building was not an intentional change I made. However it makes sense to do things that way when the group is small and so for any group with 2 or less units (includes the leader) the leader will now search the building too.